+++ src/app/shared/shared.module.ts
@@ -2,12 +2,13 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { MouseHoverStyleDirective } from './directives/mouse-hover-style.directive';
 import { ChangeStyleDirective } from './directives/change-style.directive';
+import { OrderByPipe } from './pipes/order-by.pipe';
 
 
 // Так не будет дублирования в разных массивах одних и тех же элементов
-const directives = [MouseHoverStyleDirective, ChangeStyleDirective];
+const directives = [MouseHoverStyleDirective, ChangeStyleDirective, OrderByPipe];
 @NgModule({
-  declarations: [...directives, ChangeStyleDirective],
+  declarations: [...directives],
   imports: [
     CommonModule
   ],